Rename GameEntity to Game
